Refactor namespace of Persistence, Reference, Join classes. Adds Model->addCalculatedField() method, $_default_seed_addRef property for easier extending. Split Reference getModel into multiple methods for easier extending (protected methods addToPersistence() and getDefaultPersistence()). Adds comments, more strict method parameters etc.
