+++ src/Persistence.php
@@ -5,7 +5,7 @@
 namespace atk4\data;
 
 /**
- * Class description??
+ * Persistence class.
  */
 class Persistence
 {
@@ -56,7 +56,7 @@ class Persistence
             case 'dumper':
             case 'counter':
             case 'sqlite':
-                $db = new Persistence_SQL($dsn['dsn'], $dsn['user'], $dsn['pass'], $args);
+                $db = new \atk4\data\Persistence\SQL($dsn['dsn'], $dsn['user'], $dsn['pass'], $args);
                 $db->driver = $driver;
 
                 return $db;
