+++ tests/JoinTest.php
@@ -3,7 +3,7 @@
 namespace atk4\data\tests;
 
 use atk4\data\Model;
-use atk4\data\Persistence_Array;
+use atk4\data\Persistence\Array_ as Persistence_Array;
 
 /**
  * @coversDefaultClass \atk4\data\Model
