+++ tests/RandomTest.php
@@ -4,7 +4,7 @@ namespace atk4\data\tests;
 
 use atk4\data\Model;
 use atk4\data\Persistence;
-use atk4\data\Persistence_SQL;
+use atk4\data\Persistence\SQL as Persistence_SQL;
 
 class Model_Rate extends \atk4\data\Model
 {
