+++ tests/SerializeTest.php
@@ -3,7 +3,7 @@
 namespace atk4\data\tests;
 
 use atk4\data\Model;
-use atk4\data\Persistence_SQL;
+use atk4\data\Persistence\SQL as Persistence_SQL;
 
 class SerializeTest extends \atk4\schema\PHPUnit_SchemaTestCase
 {
