+++ tests/StaticPersistenceTest.php
@@ -2,6 +2,9 @@
 
 namespace atk4\data\tests;
 
+use atk4\data\Model;
+use atk4\data\Persistence\Static_ as Persistence_Static;
+
 /**
  * @coversDefaultClass \atk4\data\Model
  */
@@ -12,23 +15,23 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
      */
     public function testBasicStatic()
     {
-        $p = new \atk4\data\Persistence_Static(['hello', 'world']);
+        $p = new Persistence_Static(['hello', 'world']);
 
         // default title field
-        $m = new \atk4\data\Model($p);
+        $m = new Model($p);
         $m->load(1);
         $this->assertEquals('world', $m['name']);
 
         // custom title field and try loading from same static twice
-        $m = new \atk4\data\Model($p); //, ['title_field' => 'foo']);
+        $m = new Model($p); //, ['title_field' => 'foo']);
         $m->load(1);
         $this->assertEquals('world', $m['name']); // still 'name' here not 'foo'
     }
 
     public function testArrayOfArrays()
     {
-        $p = new \atk4\data\Persistence_Static([['hello', 'xx', true], ['world', 'xy', false]]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['hello', 'xx', true], ['world', 'xy', false]]);
+        $m = new Model($p);
 
         $m->load(1);
 
@@ -39,8 +42,8 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
 
     public function testArrayOfHashes()
     {
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello'], ['foo'=>'world']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['foo'=>'hello'], ['foo'=>'world']]);
+        $m = new Model($p);
 
         $m->load(1);
 
@@ -49,8 +52,8 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
 
     public function testIDArg()
     {
-        $p = new \atk4\data\Persistence_Static([['id'=>20, 'foo'=>'hello'], ['id'=>21, 'foo'=>'world']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['id'=>20, 'foo'=>'hello'], ['id'=>21, 'foo'=>'world']]);
+        $m = new Model($p);
 
         $m->load(21);
 
@@ -59,8 +62,8 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
 
     public function testIDKey()
     {
-        $p = new \atk4\data\Persistence_Static([20=>['foo'=>'hello'], 21=>['foo'=>'world']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([20=>['foo'=>'hello'], 21=>['foo'=>'world']]);
+        $m = new Model($p);
 
         $m->load(21);
 
@@ -69,8 +72,8 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
 
     public function testEmpty()
     {
-        $p = new \atk4\data\Persistence_Static([]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([]);
+        $m = new Model($p);
 
         $m->tryLoadAny();
 
@@ -79,34 +82,34 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
 
     public function testCustomField()
     {
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello'], ['foo'=>'world']]);
+        $p = new Persistence_Static([['foo'=>'hello'], ['foo'=>'world']]);
         $m = new StaticPersistenceModel($p);
 
         $this->assertEquals('custom field', $m->getElement('foo')->caption);
 
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello', 'bar'=>'world']]);
+        $p = new Persistence_Static([['foo'=>'hello', 'bar'=>'world']]);
         $m = new StaticPersistenceModel($p);
         $this->assertEquals('foo', $m->title_field);
     }
 
     public function testTitleOrName()
     {
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello', 'bar'=>'world']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['foo'=>'hello', 'bar'=>'world']]);
+        $m = new Model($p);
         $this->assertEquals('foo', $m->title_field);
 
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello', 'name'=>'x']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['foo'=>'hello', 'name'=>'x']]);
+        $m = new Model($p);
         $this->assertEquals('name', $m->title_field);
 
-        $p = new \atk4\data\Persistence_Static([['foo'=>'hello', 'title'=>'x']]);
-        $m = new \atk4\data\Model($p);
+        $p = new Persistence_Static([['foo'=>'hello', 'title'=>'x']]);
+        $m = new Model($p);
         $this->assertEquals('title', $m->title_field);
     }
 
     public function testFieldTypes()
     {
-        $p = new \atk4\data\Persistence_Static([[
+        $p = new Persistence_Static([[
             'name'        => 'hello',
             'test_int'    => 123,
             'test_float'  => 123.45,
@@ -117,7 +120,7 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
             'test_str_2'  => '123',
             'test_str_3'  => '123.45',
         ]]);
-        $m = new \atk4\data\Model($p);
+        $m = new Model($p);
 
         $this->assertEquals('integer', $m->getElement('test_int')->type);
         $this->assertEquals('float', $m->getElement('test_float')->type);
@@ -133,7 +136,7 @@ class StaticPersistenceTest extends \atk4\core\PHPUnit_AgileTestCase
     }
 }
 
-class StaticPersistenceModel extends \atk4\data\Model
+class StaticPersistenceModel extends Model
 {
     public $title_field = 'foo';
